fix(ride-model): make ride model work

+++ ride-service/core/src/main/kotlin/it/unibo/sap/ass02/domain/model/P2d.kt
@@ -1,22 +1,20 @@
 package it.unibo.sap.ass02.domain.model
 
-import it.unibo.sap.ass02.domain.model.V2d.Companion.fromCoord
-
 interface P2d {
-    val x: Int
-    val y: Int
+    val x: Double
+    val y: Double
 
     operator fun plus(other: P2d): P2d = fromCoord(x + other.x, y + other.y)
 
-    operator fun plus(other: V2d): P2d = fromCoord((x + other.x).toInt(), (y + other.y).toInt())
+    operator fun plus(other: V2d): P2d = fromCoord(x + other.x, y + other.y)
 
     companion object {
         fun fromCoord(
-            x: Int,
-            y: Int,
+            x: Double,
+            y: Double,
         ) = object : P2d {
-            override val x: Int = x
-            override val y: Int = y
+            override val x: Double = x
+            override val y: Double = y
         }
     }
 }
